# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Driver API",
        "title": "Maratona Full Cycle Driver API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sidarta Silva",
            "url": "http://www.fullcycle.com.br",
            "email": "atendimento@fullcycle.com.br"
        },
        "license": {
            "name": "Full Cycle License",
            "url": "http://www.fullcycle.com.br"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/drivers": {
            "get": {
                "description": "Get all drivers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "List drivers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Driver"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/drivers/{id}": {
            "get": {
                "description": "Get a driver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Get a driver",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Driver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Driver": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Sidarta S"
                },
                "uuid": {
                    "type": "string",
                    "example": "127a2285-0fd8-4df9-b4eb-485292f3e345"
                }
            }
        }
    }
}